Remove xdebug package name
xdebug is debugging for any custom language, so Elixir will only ever
have xdebug, so xdebug in the name is unnecessary.
