+++ src/org/elixir_lang/beam/term/Inspect.kt
@@ -2,10 +2,10 @@ package org.elixir_lang.beam.term
 
 import com.ericsson.otp.erlang.OtpErlangObject
 import org.elixir_lang.beam.chunk.XValueRenderer
-import org.elixir_lang.debugger.XValuePresentation
+import org.elixir_lang.debugger.stack_frame.value.Presentation
 
 fun inspect(term: OtpErlangObject): String =
     XValueRenderer().let { renderer ->
-        XValuePresentation(term).renderValue(renderer)
+        Presentation(term).renderValue(renderer)
         renderer.getText()
     }
