+++ src/org/elixir_lang/debugger/stack_frame/value/Atom.java
@@ -16,19 +16,12 @@
  * limitations under the License.
  */
 
-package org.elixir_lang.debugger.xdebug.xvalue;
+package org.elixir_lang.debugger.stack_frame.value;
 
-import com.ericsson.otp.erlang.OtpErlangTuple;
-import com.intellij.xdebugger.frame.XValueChildrenList;
-import org.jetbrains.annotations.NotNull;
+import com.ericsson.otp.erlang.OtpErlangAtom;
 
-class Tuple extends ArrayBase<OtpErlangTuple> {
-  Tuple(@NotNull OtpErlangTuple value) {
-    super(value, value.arity());
-  }
-
-  @Override
-  protected void computeChild(@NotNull XValueChildrenList children, int childIdx) {
-    addIndexedChild(children, getValue().elementAt(childIdx), childIdx);
+class Atom extends PrimitiveBase<OtpErlangAtom> {
+  Atom(OtpErlangAtom value) {
+    super(value);
   }
 }
