+++ src/org/elixir_lang/debugger/stack_frame/value/Numeric.java
@@ -16,24 +16,24 @@
  * limitations under the License.
  */
 
-package org.elixir_lang.debugger.xdebug.xvalue;
+package org.elixir_lang.debugger.stack_frame.value;
 
-import com.ericsson.otp.erlang.OtpErlangPid;
-import com.intellij.xdebugger.frame.XCompositeNode;
-import com.intellij.xdebugger.frame.XValueChildrenList;
+import com.ericsson.otp.erlang.OtpErlangObject;
+import com.intellij.xdebugger.frame.XValueNode;
+import com.intellij.xdebugger.frame.XValuePlace;
+import com.intellij.xdebugger.frame.presentation.XNumericValuePresentation;
+import com.intellij.xdebugger.frame.presentation.XValuePresentation;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
-class Pid extends Base<OtpErlangPid> {
-  Pid(OtpErlangPid value) {
-    super(value, 3);
+class Numeric extends PrimitiveBase<OtpErlangObject> {
+  Numeric(OtpErlangObject value) {
+    super(value);
   }
 
+  @Nullable
   @Override
-  public void computeChildren(@NotNull XCompositeNode node) {
-    XValueChildrenList childrenList = new XValueChildrenList(3);
-    addNamedChild(childrenList, getValue().node(), "node");
-    addNamedChild(childrenList, getValue().id(), "id");
-    addNamedChild(childrenList, getValue().serial(), "serial");
-    node.addChildren(childrenList, true);
+  protected XValuePresentation getPresentation(@NotNull XValueNode node, @NotNull XValuePlace place) {
+    return new XNumericValuePresentation(getValue().toString());
   }
 }
