+++ src/org/elixir_lang/debugger/stack_frame/value/Pid.java
@@ -16,26 +16,24 @@
  * limitations under the License.
  */
 
-package org.elixir_lang.debugger.xdebug.xvalue;
+package org.elixir_lang.debugger.stack_frame.value;
 
-import com.ericsson.otp.erlang.OtpErlangRef;
+import com.ericsson.otp.erlang.OtpErlangPid;
 import com.intellij.xdebugger.frame.XCompositeNode;
 import com.intellij.xdebugger.frame.XValueChildrenList;
 import org.jetbrains.annotations.NotNull;
 
-public class Ref extends Base<OtpErlangRef> {
-  Ref(@NotNull OtpErlangRef value) {
-    super(value, 1 + value.ids().length);
+class Pid extends Base<OtpErlangPid> {
+  Pid(OtpErlangPid value) {
+    super(value, 3);
   }
 
   @Override
   public void computeChildren(@NotNull XCompositeNode node) {
-    int[] ids = getValue().ids();
-    XValueChildrenList children = new XValueChildrenList(1 + ids.length);
-    addNamedChild(children, getValue().node(), "node");
-    for (int i = 0; i < ids.length; i++) {
-      addNamedChild(children, ids[i], "id" + i);
-    }
-    node.addChildren(children, true);
+    XValueChildrenList childrenList = new XValueChildrenList(3);
+    addNamedChild(childrenList, getValue().node(), "node");
+    addNamedChild(childrenList, getValue().id(), "id");
+    addNamedChild(childrenList, getValue().serial(), "serial");
+    node.addChildren(childrenList, true);
   }
 }
