+++ src/org/elixir_lang/debugger/stack_frame/value/Port.java
@@ -16,19 +16,23 @@
  * limitations under the License.
  */
 
-package org.elixir_lang.debugger.xdebug.xvalue;
+package org.elixir_lang.debugger.stack_frame.value;
 
-import com.ericsson.otp.erlang.OtpErlangBitstr;
+import com.ericsson.otp.erlang.OtpErlangPort;
+import com.intellij.xdebugger.frame.XCompositeNode;
 import com.intellij.xdebugger.frame.XValueChildrenList;
 import org.jetbrains.annotations.NotNull;
 
-class BitString extends ArrayBase<OtpErlangBitstr> {
-  BitString(@NotNull OtpErlangBitstr value) {
-    super(value, value.binaryValue().length);
+public class Port extends Base<OtpErlangPort> {
+  Port(OtpErlangPort value) {
+    super(value, 2);
   }
 
   @Override
-  protected void computeChild(@NotNull XValueChildrenList children, int childIdx) {
-    addIndexedChild(children, getValue().binaryValue()[childIdx] & 0xFF, childIdx);
+  public void computeChildren(@NotNull XCompositeNode node) {
+    XValueChildrenList childrenList = new XValueChildrenList(2);
+    addNamedChild(childrenList, getValue().node(), "node");
+    addNamedChild(childrenList, getValue().id(), "id");
+    node.addChildren(childrenList, true);
   }
 }
